Simplified Infoview interaction; focus returns to Lean file when opening infoview; Removed old goal view
The states of the info view are now: display goals, display list or stop updating. When updating is stopped, the user activates it with either display goals or disply list. These are now commands on the Lean file (not the info view anymore). The command sequence `Ctrl+Shift+Enter` opens the goal display in the infoview. I removed the old goal view.